made griLayout functional, added RenderedGridElements in GridLayout
